requirements_updated.txt
streamlit
pandas
numpy
requests
geopy
pydeck
scikit-learn
openai
python-dotenv
Flask
flask-cors
python-dotenv
geopy
streamlit
